+++ beacon-chain/rpc/prysm/v1alpha1/validator/aggregator.go
@@ -60,7 +60,7 @@ func (vs *Server) SubmitAggregateSelectionProof(ctx context.Context, req *ethpb.
 	}
 
 	if err := vs.AttPool.AggregateUnaggregatedAttestationsBySlotIndex(ctx, req.Slot, req.CommitteeIndex); err != nil {
-		return nil, status.Errorf(codes.Internal, "Could not aggregate unaggregated attestations")
+		return nil, status.Errorf(codes.Internal, "Could not aggregate unaggregated Attestations")
 	}
 	aggregatedAtts := vs.AttPool.AggregatedAttestationsBySlotIndex(ctx, req.Slot, req.CommitteeIndex)
