+++ beacon-chain/rpc/prysm/v1alpha1/validator/proposer.go
@@ -12,7 +12,6 @@ import (
 	fastssz "github.com/ferranbt/fastssz"
 	"github.com/pkg/errors"
 	types "github.com/prysmaticlabs/eth2-types"
-	"github.com/prysmaticlabs/go-bitfield"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/feed"
 	blockfeed "github.com/prysmaticlabs/prysm/beacon-chain/core/feed/block"
@@ -22,7 +21,6 @@ import (
 	iface "github.com/prysmaticlabs/prysm/beacon-chain/state/interface"
 	dbpb "github.com/prysmaticlabs/prysm/proto/beacon/db"
 	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
-	prysmv2 "github.com/prysmaticlabs/prysm/proto/prysm/v2"
 	attaggregation "github.com/prysmaticlabs/prysm/shared/aggregation/attestations"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/featureconfig"
@@ -52,17 +50,17 @@ type eth1DataAggregatedVote struct {
 	votes int
 }
 
-// Data required to create a beacon block.
-type blockData struct {
-	parentRoot        []byte
-	graffiti          [32]byte
-	proposerIdx       types.ValidatorIndex
-	eth1Data          *ethpb.Eth1Data
-	deposits          []*ethpb.Deposit
-	attestations      []*ethpb.Attestation
-	proposerSlashings []*ethpb.ProposerSlashing
-	attesterSlashings []*ethpb.AttesterSlashing
-	voluntaryExits    []*ethpb.SignedVoluntaryExit
+// BlockData required to create a beacon block.
+type BlockData struct {
+	ParentRoot        []byte
+	Graffiti          [32]byte
+	ProposerIdx       types.ValidatorIndex
+	Eth1Data          *ethpb.Eth1Data
+	Deposits          []*ethpb.Deposit
+	Attestations      []*ethpb.Attestation
+	ProposerSlashings []*ethpb.ProposerSlashing
+	AttesterSlashings []*ethpb.AttesterSlashing
+	VoluntaryExits    []*ethpb.SignedVoluntaryExit
 }
 
 // GetBlock is called by a proposer during its assigned slot to request a block to sign
@@ -72,7 +70,7 @@ func (vs *Server) GetBlock(ctx context.Context, req *ethpb.BlockRequest) (*ethpb
 	defer span.End()
 	span.AddAttributes(trace.Int64Attribute("slot", int64(req.Slot)))
 
-	blkData, err := vs.buildBlockData(ctx, req)
+	blkData, err := vs.BuildBlockData(ctx, req)
 	if err != nil {
 		return nil, err
 	}
@@ -82,23 +80,27 @@ func (vs *Server) GetBlock(ctx context.Context, req *ethpb.BlockRequest) (*ethpb
 
 	blk := &ethpb.BeaconBlock{
 		Slot:          req.Slot,
-		ParentRoot:    blkData.parentRoot,
+		ParentRoot:    blkData.ParentRoot,
 		StateRoot:     stateRoot,
-		ProposerIndex: blkData.proposerIdx,
+		ProposerIndex: blkData.ProposerIdx,
 		Body: &ethpb.BeaconBlockBody{
-			Eth1Data:          blkData.eth1Data,
-			Deposits:          blkData.deposits,
-			Attestations:      blkData.attestations,
+			Eth1Data:          blkData.Eth1Data,
+			Deposits:          blkData.Deposits,
+			Attestations:      blkData.Attestations,
 			RandaoReveal:      req.RandaoReveal,
-			ProposerSlashings: blkData.proposerSlashings,
-			AttesterSlashings: blkData.attesterSlashings,
-			VoluntaryExits:    blkData.voluntaryExits,
-			Graffiti:          blkData.graffiti[:],
+			ProposerSlashings: blkData.ProposerSlashings,
+			AttesterSlashings: blkData.AttesterSlashings,
+			VoluntaryExits:    blkData.VoluntaryExits,
+			Graffiti:          blkData.Graffiti[:],
 		},
 	}
 
 	// Compute state root with the newly constructed block.
-	stateRoot, err = vs.computeStateRoot(ctx, interfaces.WrappedPhase0SignedBeaconBlock(&ethpb.SignedBeaconBlock{Block: blk, Signature: make([]byte, 96)}))
+	stateRoot, err = vs.ComputeStateRoot(
+		ctx, interfaces.WrappedPhase0SignedBeaconBlock(
+			&ethpb.SignedBeaconBlock{Block: blk, Signature: make([]byte, 96)},
+		),
+	)
 	if err != nil {
 		interop.WriteBlockToDisk(interfaces.WrappedPhase0SignedBeaconBlock(&ethpb.SignedBeaconBlock{Block: blk}), true /*failed*/)
 		return nil, status.Errorf(codes.Internal, "Could not compute state root: %v", err)
@@ -108,67 +110,9 @@ func (vs *Server) GetBlock(ctx context.Context, req *ethpb.BlockRequest) (*ethpb
 	return blk, nil
 }
 
-// GetBlockV2 is called by a proposer during its assigned slot to request a block to sign
-// by passing in the slot and the signed randao reveal of the slot. This is used by a validator
-// after the altair fork epoch has been encountered.
-func (vs *Server) GetBlockV2(ctx context.Context, req *ethpb.BlockRequest) (*prysmv2.BeaconBlock, error) {
-	ctx, span := trace.StartSpan(ctx, "ProposerServer.GetBlockV2")
-	defer span.End()
-	span.AddAttributes(trace.Int64Attribute("slot", int64(req.Slot)))
-
-	blkData, err := vs.buildBlockData(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-
-	// Use zero hash as stub for state root to compute later.
-	stateRoot := params.BeaconConfig().ZeroHash[:]
-
-	// Ugly hack to allow this to compile both for mainnet
-	// and minimal configs.
-	mockAgg := &prysmv2.SyncAggregate{SyncCommitteeBits: []byte{}}
-	var bVector []byte
-	if mockAgg.SyncCommitteeBits.Len() == 512 {
-		bVector = bitfield.NewBitvector512()
-	} else {
-		bVector = bitfield.NewBitvector32()
-	}
-	infiniteSignature := [96]byte{0xC0}
-	blk := &prysmv2.BeaconBlock{
-		Slot:          req.Slot,
-		ParentRoot:    blkData.parentRoot,
-		StateRoot:     stateRoot,
-		ProposerIndex: blkData.proposerIdx,
-		Body: &prysmv2.BeaconBlockBody{
-			Eth1Data:          blkData.eth1Data,
-			Deposits:          blkData.deposits,
-			Attestations:      blkData.attestations,
-			RandaoReveal:      req.RandaoReveal,
-			ProposerSlashings: blkData.proposerSlashings,
-			AttesterSlashings: blkData.attesterSlashings,
-			VoluntaryExits:    blkData.voluntaryExits,
-			Graffiti:          blkData.graffiti[:],
-			// TODO: Add in actual aggregates
-			SyncAggregate: &prysmv2.SyncAggregate{
-				SyncCommitteeBits:      bVector,
-				SyncCommitteeSignature: infiniteSignature[:],
-			},
-		},
-	}
-	// Compute state root with the newly constructed block.
-	stateRoot, err = vs.computeStateRoot(ctx, interfaces.WrappedAltairSignedBeaconBlock(&prysmv2.SignedBeaconBlock{Block: blk, Signature: make([]byte, 96)}))
-	if err != nil {
-		interop.WriteBlockToDisk(interfaces.WrappedAltairSignedBeaconBlock(&prysmv2.SignedBeaconBlock{Block: blk}), true /*failed*/)
-		return nil, status.Errorf(codes.Internal, "Could not compute state root: %v", err)
-	}
-	blk.StateRoot = stateRoot
-
-	return blk, nil
-}
-
-// builds the relevant data for creating a new beacon block, so that this method can be shared across forks.
-func (vs *Server) buildBlockData(ctx context.Context, req *ethpb.BlockRequest) (*blockData, error) {
-	ctx, span := trace.StartSpan(ctx, "ProposerServer.buildBlockData")
+// BuildBlockData for creating a new beacon block, so that this method can be shared across forks.
+func (vs *Server) BuildBlockData(ctx context.Context, req *ethpb.BlockRequest) (*BlockData, error) {
+	ctx, span := trace.StartSpan(ctx, "ProposerServer.BuildBlockData")
 	defer span.End()
 
 	if vs.SyncChecker.Syncing() {
@@ -203,16 +147,16 @@ func (vs *Server) buildBlockData(ctx context.Context, req *ethpb.BlockRequest) (
 		return nil, status.Errorf(codes.Internal, "Could not get ETH1 data: %v", err)
 	}
 
-	// Pack ETH1 deposits which have not been included in the beacon chain.
+	// Pack ETH1 Deposits which have not been included in the beacon chain.
 	deposits, err := vs.deposits(ctx, head, eth1Data)
 	if err != nil {
-		return nil, status.Errorf(codes.Internal, "Could not get ETH1 deposits: %v", err)
+		return nil, status.Errorf(codes.Internal, "Could not get ETH1 Deposits: %v", err)
 	}
 
-	// Pack aggregated attestations which have not been included in the beacon chain.
+	// Pack aggregated Attestations which have not been included in the beacon chain.
 	atts, err := vs.packAttestations(ctx, head)
 	if err != nil {
-		return nil, status.Errorf(codes.Internal, "Could not get attestations to pack into block: %v", err)
+		return nil, status.Errorf(codes.Internal, "Could not get Attestations to pack into block: %v", err)
 	}
 
 	graffiti := bytesutil.ToBytes32(req.Graffiti)
@@ -223,16 +167,16 @@ func (vs *Server) buildBlockData(ctx context.Context, req *ethpb.BlockRequest) (
 		return nil, status.Errorf(codes.Internal, "Could not calculate proposer index %v", err)
 	}
 
-	return &blockData{
-		parentRoot:        parentRoot,
-		graffiti:          graffiti,
-		proposerIdx:       idx,
-		eth1Data:          eth1Data,
-		deposits:          deposits,
-		attestations:      atts,
-		proposerSlashings: vs.SlashingsPool.PendingProposerSlashings(ctx, head, false /*noLimit*/),
-		attesterSlashings: vs.SlashingsPool.PendingAttesterSlashings(ctx, head, false /*noLimit*/),
-		voluntaryExits:    vs.ExitPool.PendingExits(head, req.Slot, false /*noLimit*/),
+	return &BlockData{
+		ParentRoot:        parentRoot,
+		Graffiti:          graffiti,
+		ProposerIdx:       idx,
+		Eth1Data:          eth1Data,
+		Deposits:          deposits,
+		Attestations:      atts,
+		ProposerSlashings: vs.SlashingsPool.PendingProposerSlashings(ctx, head, false /*noLimit*/),
+		AttesterSlashings: vs.SlashingsPool.PendingAttesterSlashings(ctx, head, false /*noLimit*/),
+		VoluntaryExits:    vs.ExitPool.PendingExits(head, req.Slot, false /*noLimit*/),
 	}, nil
 }
 
@@ -240,19 +184,11 @@ func (vs *Server) buildBlockData(ctx context.Context, req *ethpb.BlockRequest) (
 // to get it processed by the beacon node as the canonical head.
 func (vs *Server) ProposeBlock(ctx context.Context, rBlk *ethpb.SignedBeaconBlock) (*ethpb.ProposeResponse, error) {
 	blk := interfaces.WrappedPhase0SignedBeaconBlock(rBlk)
-	return vs.proposeBlock(ctx, blk)
+	return vs.ProposeBlockGeneric(ctx, blk)
 }
 
-// ProposeBlockV2 is called by a proposer during its assigned slot to create a block in an attempt
-// to get it processed by the beacon node as the canonical head.
-func (vs *Server) ProposeBlockV2(ctx context.Context, rBlk *prysmv2.SignedBeaconBlock) (*ethpb.ProposeResponse, error) {
-	blk := interfaces.WrappedAltairSignedBeaconBlock(rBlk)
-	return vs.proposeBlock(ctx, blk)
-}
-
-// propose block performs the core post-block creation actions once a block proposal is
-// received.
-func (vs *Server) proposeBlock(ctx context.Context, blk interfaces.SignedBeaconBlock) (*ethpb.ProposeResponse, error) {
+// ProposeBlockGeneric performs the core post-block creation actions once a block proposal is received.
+func (vs *Server) ProposeBlockGeneric(ctx context.Context, blk interfaces.SignedBeaconBlock) (*ethpb.ProposeResponse, error) {
 	ctx, span := trace.StartSpan(ctx, "ProposerServer.proposeBlock")
 	defer span.End()
 	root, err := blk.Block().HashTreeRoot()
@@ -475,9 +411,9 @@ func (vs *Server) randomETH1DataVote(ctx context.Context) (*ethpb.Eth1Data, erro
 	}, nil
 }
 
-// computeStateRoot computes the state root after a block has been processed through a state transition and
+// ComputeStateRoot computes the state root after a block has been processed through a state transition and
 // returns it to the validator client.
-func (vs *Server) computeStateRoot(ctx context.Context, block interfaces.SignedBeaconBlock) ([]byte, error) {
+func (vs *Server) ComputeStateRoot(ctx context.Context, block interfaces.SignedBeaconBlock) ([]byte, error) {
 	beaconState, err := vs.StateGen.StateByRoot(ctx, bytesutil.ToBytes32(block.Block().ParentRoot()))
 	if err != nil {
 		return nil, errors.Wrap(err, "could not retrieve beacon state")
@@ -505,14 +441,14 @@ func (vs *Server) deposits(
 	beaconState iface.BeaconState,
 	currentVote *ethpb.Eth1Data,
 ) ([]*ethpb.Deposit, error) {
-	ctx, span := trace.StartSpan(ctx, "ProposerServer.deposits")
+	ctx, span := trace.StartSpan(ctx, "ProposerServer.Deposits")
 	defer span.End()
 
 	if vs.MockEth1Votes || !vs.Eth1InfoFetcher.IsConnectedToETH1() {
 		return []*ethpb.Deposit{}, nil
 	}
-	// Need to fetch if the deposits up to the state's latest eth 1 data matches
-	// the number of all deposits in this RPC call. If not, then we return nil.
+	// Need to fetch if the Deposits up to the state's latest eth 1 data matches
+	// the number of all Deposits in this RPC call. If not, then we return nil.
 	canonicalEth1Data, canonicalEth1DataHeight, err := vs.canonicalEth1Data(ctx, beaconState, currentVote)
 	if err != nil {
 		return nil, err
@@ -523,7 +459,7 @@ func (vs *Server) deposits(
 		return []*ethpb.Deposit{}, nil
 	}
 
-	// If there are no pending deposits, exit early.
+	// If there are no pending Deposits, exit early.
 	allPendingContainers := vs.PendingDepositsFetcher.PendingContainers(ctx, canonicalEth1DataHeight)
 	if len(allPendingContainers) == 0 {
 		return []*ethpb.Deposit{}, nil
@@ -535,7 +471,7 @@ func (vs *Server) deposits(
 	}
 
 	// Deposits need to be received in order of merkle index root, so this has to make sure
-	// deposits are sorted from lowest to highest.
+	// Deposits are sorted from lowest to highest.
 	var pendingDeps []*dbpb.DepositContainer
 	for _, dep := range allPendingContainers {
 		if uint64(dep.Index) >= beaconState.Eth1DepositIndex() && uint64(dep.Index) < canonicalEth1Data.DepositCount {
@@ -544,7 +480,7 @@ func (vs *Server) deposits(
 	}
 
 	for i := range pendingDeps {
-		// Don't construct merkle proof if the number of deposits is more than max allowed in block.
+		// Don't construct merkle proof if the number of Deposits is more than max allowed in block.
 		if uint64(i) == params.BeaconConfig().MaxDeposits {
 			break
 		}
@@ -553,7 +489,7 @@ func (vs *Server) deposits(
 			return nil, err
 		}
 	}
-	// Limit the return of pending deposits to not be more than max deposits allowed in block.
+	// Limit the return of pending Deposits to not be more than max Deposits allowed in block.
 	var pendingDeposits []*ethpb.Deposit
 	for i := uint64(0); i < uint64(len(pendingDeps)) && i < params.BeaconConfig().MaxDeposits; i++ {
 		pendingDeposits = append(pendingDeposits, pendingDeps[i].Deposit)
@@ -675,7 +611,7 @@ func (vs *Server) defaultEth1DataResponse(ctx context.Context, currentHeight *bi
 	if err != nil {
 		return nil, errors.Wrap(err, "could not fetch ETH1_FOLLOW_DISTANCE ancestor")
 	}
-	// Fetch all historical deposits up to an ancestor height.
+	// Fetch all historical Deposits up to an ancestor height.
 	depositsTillHeight, depositRoot := vs.DepositFetcher.DepositsNumberAndRootAtHeight(ctx, ancestorHeight)
 	if depositsTillHeight == 0 {
 		return vs.ChainStartFetcher.ChainStartEth1Data(), nil
@@ -692,7 +628,7 @@ func (vs *Server) defaultEth1DataResponse(ctx context.Context, currentHeight *bi
 	}, nil
 }
 
-// This filters the input attestations to return a list of valid attestations to be packaged inside a beacon block.
+// This filters the input Attestations to return a list of valid Attestations to be packaged inside a beacon block.
 func (vs *Server) filterAttestationsForBlockInclusion(ctx context.Context, st iface.BeaconState, atts []*ethpb.Attestation) ([]*ethpb.Attestation, error) {
 	ctx, span := trace.StartSpan(ctx, "ProposerServer.filterAttestationsForBlockInclusion")
 	defer span.End()
@@ -704,7 +640,7 @@ func (vs *Server) filterAttestationsForBlockInclusion(ctx context.Context, st if
 	return validAtts.dedup().sortByProfitability().limitToMaxAttestations(), nil
 }
 
-// The input attestations are processed and seen by the node, this deletes them from pool
+// The input Attestations are processed and seen by the node, this deletes them from pool
 // so proposers don't include them in a block for the future.
 func (vs *Server) deleteAttsInPool(ctx context.Context, atts []*ethpb.Attestation) error {
 	ctx, span := trace.StartSpan(ctx, "ProposerServer.deleteAttsInPool")
@@ -733,7 +669,7 @@ func constructMerkleProof(trie *trieutil.SparseMerkleTrie, index int, deposit *e
 		return nil, errors.Wrapf(err, "could not generate merkle proof for deposit at index %d", index)
 	}
 	// For every deposit, we construct a Merkle proof using the powchain service's
-	// in-memory deposits trie, which is updated only once the state's LatestETH1Data
+	// in-memory Deposits trie, which is updated only once the state's LatestETH1Data
 	// property changes during a state transition after a voting period.
 	deposit.Proof = proof
 	return deposit, nil
@@ -746,19 +682,19 @@ func (vs *Server) packAttestations(ctx context.Context, latestState iface.Beacon
 	atts := vs.AttPool.AggregatedAttestations()
 	atts, err := vs.filterAttestationsForBlockInclusion(ctx, latestState, atts)
 	if err != nil {
-		return nil, errors.Wrap(err, "could not filter attestations")
+		return nil, errors.Wrap(err, "could not filter Attestations")
 	}
 
-	// If there is any room left in the block, consider unaggregated attestations as well.
+	// If there is any room left in the block, consider unaggregated Attestations as well.
 	numAtts := uint64(len(atts))
 	if numAtts < params.BeaconConfig().MaxAttestations {
 		uAtts, err := vs.AttPool.UnaggregatedAttestations()
 		if err != nil {
-			return nil, errors.Wrap(err, "could not get unaggregated attestations")
+			return nil, errors.Wrap(err, "could not get unaggregated Attestations")
 		}
 		uAtts, err = vs.filterAttestationsForBlockInclusion(ctx, latestState, uAtts)
 		if err != nil {
-			return nil, errors.Wrap(err, "could not filter attestations")
+			return nil, errors.Wrap(err, "could not filter Attestations")
 		}
 		atts = append(atts, uAtts...)
