+++ beacon-chain/rpc/prysm/v1alpha1/validator/server.go
@@ -1,5 +1,5 @@
 // Package validator defines a gRPC validator service implementation, providing
-// critical endpoints for validator clients to submit blocks/attestations to the
+// critical endpoints for validator clients to submit blocks/Attestations to the
 // beacon node, receive assignments, and more.
 package validator
