+++ beacon-chain/rpc/prysm/v2/validator/sync_committee.go
